+++ src/cofd/parser.rs
@@ -1,8 +1,8 @@
-use crate::cofd::dice::{Amount, DicePool, DicePoolModifiers, DicePoolQuality, Element, Operator};
+use crate::cofd::dice::{DicePool, DicePoolModifiers, DicePoolQuality};
 use crate::error::BotError;
-use combine::error::StringStreamError;
-use combine::parser::char::{digit, letter, spaces, string};
-use combine::{choice, count, many, many1, one_of, Parser};
+use crate::parser::{parse_amounts, DiceParsingError};
+use combine::parser::char::{digit, spaces, string};
+use combine::{choice, count, many1, one_of, Parser};
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 enum ParsedInfo {
@@ -10,37 +10,7 @@ enum ParsedInfo {
     ExceptionalOn(i32),
 }
 
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum DiceParsingError {
-    InvalidAmount,
-    InvalidModifiers,
-    UnconsumedInput,
-}
-
-impl std::fmt::Display for DiceParsingError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.as_str())
-    }
-}
-
-impl std::error::Error for DiceParsingError {
-    fn description(&self) -> &str {
-        self.as_str()
-    }
-}
-
-impl DiceParsingError {
-    fn as_str(&self) -> &str {
-        use self::DiceParsingError::*;
-        match *self {
-            InvalidAmount => "invalid amount of dice",
-            InvalidModifiers => "dice pool modifiers not specified properly",
-            UnconsumedInput => "extraneous input detected",
-        }
-    }
-}
-
-pub fn parse_modifiers(input: &str) -> Result<DicePoolModifiers, BotError> {
+pub fn parse_modifiers(input: &str) -> Result<DicePoolModifiers, DiceParsingError> {
     if input.len() == 0 {
         return Ok(DicePoolModifiers::default());
     }
@@ -77,13 +47,11 @@ pub fn parse_modifiers(input: &str) -> Result<DicePoolModifiers, BotError> {
     if rest.len() == 0 {
         convert_to_info(&result)
     } else {
-        Err(BotError::DiceParsingError(
-            DiceParsingError::UnconsumedInput,
-        ))
+        Err(DiceParsingError::UnconsumedInput)
     }
 }
 
-fn convert_to_info(parsed: &Vec<ParsedInfo>) -> Result<DicePoolModifiers, BotError> {
+fn convert_to_info(parsed: &Vec<ParsedInfo>) -> Result<DicePoolModifiers, DiceParsingError> {
     use ParsedInfo::*;
     if parsed.len() == 0 {
         Ok(DicePoolModifiers::default())
@@ -110,77 +78,6 @@ fn convert_to_info(parsed: &Vec<ParsedInfo>) -> Result<DicePoolModifiers, BotErr
     }
 }
 
-/// Parse dice pool amounts into elements coupled with operators,
-/// where an operator is "+" or "-", and an element is either a number
-/// or variable name. The first element should not have an operator,
-/// but every one after that should. Accepts expressions like "8", "10
-/// + variablename", "variablename - 3", etc.
-fn parse_pool_amount(input: &str) -> Result<Vec<Amount>, BotError> {
-    let input = input.trim();
-
-    let plus_or_minus = one_of("+-".chars());
-    let maybe_sign = plus_or_minus.map(|sign: char| match sign {
-        '+' => Operator::Plus,
-        '-' => Operator::Minus,
-        _ => Operator::Plus,
-    });
-
-    //TODO make this a macro or something
-    let first = many1(letter())
-        .or(many1(digit()))
-        .skip(spaces().silent()) //Consume any space after first amount
-        .map(|value: String| match value.parse::<i32>() {
-            Ok(num) => Amount {
-                operator: Operator::Plus,
-                element: Element::Number(num),
-            },
-            _ => Amount {
-                operator: Operator::Plus,
-                element: Element::Variable(value),
-            },
-        });
-
-    let variable_or_number =
-        many1(letter())
-            .or(many1(digit()))
-            .map(|value: String| match value.parse::<i32>() {
-                Ok(num) => Element::Number(num),
-                _ => Element::Variable(value),
-            });
-
-    let sign_and_word = maybe_sign
-        .skip(spaces().silent())
-        .and(variable_or_number)
-        .skip(spaces().silent())
-        .map(|parsed: (Operator, Element)| Amount {
-            operator: parsed.0,
-            element: parsed.1,
-        });
-
-    let rest = many(sign_and_word).map(|expr: Vec<_>| expr);
-
-    let mut parser = first.and(rest);
-
-    //Maps the found expression into a Vec of Amount instances,
-    //tacking the first one on.
-    type ParsedAmountExpr = (Amount, Vec<Amount>);
-    let (results, rest) = parser
-        .parse(input)
-        .map(|mut results: (ParsedAmountExpr, &str)| {
-            let mut amounts = vec![(results.0).0];
-            amounts.append(&mut (results.0).1);
-            (amounts, results.1)
-        })?;
-
-    if rest.len() == 0 {
-        Ok(results)
-    } else {
-        Err(BotError::DiceParsingError(
-            DiceParsingError::UnconsumedInput,
-        ))
-    }
-}
-
 pub fn parse_dice_pool(input: &str) -> Result<DicePool, BotError> {
     //The "modifiers:" part is optional. Assume amounts if no modifier
     //section found.
@@ -194,68 +91,17 @@ pub fn parse_dice_pool(input: &str) -> Result<DicePool, BotError> {
     })?;
 
     let modifiers = parse_modifiers(modifiers_str)?;
-    let amounts = parse_pool_amount(&amounts_str)?;
+    let amounts = parse_amounts(&amounts_str)?;
     Ok(DicePool::new(amounts, modifiers))
 }
 
-pub fn create_chance_die() -> Result<DicePool, StringStreamError> {
+pub fn create_chance_die() -> Result<DicePool, BotError> {
     Ok(DicePool::chance_die())
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn parse_single_number_amount_test() {
-        let result = parse_pool_amount("1");
-        assert!(result.is_ok());
-        assert_eq!(
-            result.unwrap(),
-            vec![Amount {
-                operator: Operator::Plus,
-                element: Element::Number(1)
-            }]
-        );
-
-        let result = parse_pool_amount("10");
-        assert!(result.is_ok());
-        assert_eq!(
-            result.unwrap(),
-            vec![Amount {
-                operator: Operator::Plus,
-                element: Element::Number(10)
-            }]
-        );
-    }
-
-    #[test]
-    fn parse_single_variable_amount_test() {
-        let result = parse_pool_amount("asdf");
-        assert!(result.is_ok());
-        assert_eq!(
-            result.unwrap(),
-            vec![Amount {
-                operator: Operator::Plus,
-                element: Element::Variable("asdf".to_string())
-            }]
-        );
-
-        let result = parse_pool_amount("nosis");
-        assert!(result.is_ok());
-        assert_eq!(
-            result.unwrap(),
-            vec![Amount {
-                operator: Operator::Plus,
-                element: Element::Variable("nosis".to_string())
-            }]
-        );
-    }
-
-    #[test]
-    fn parse_complex_amount_expression() {
-        assert!(parse_pool_amount("1 + myvariable - 2").is_ok());
-    }
 
     #[test]
     fn quality_test() {
@@ -289,24 +135,14 @@ mod tests {
 
         let result = parse_modifiers("b");
         assert!(result.is_err());
-        assert!(matches!(
-            result,
-            Err(BotError::DiceParsingError(
-                DiceParsingError::UnconsumedInput
-            ))
-        ));
+        assert!(matches!(result, Err(DiceParsingError::UnconsumedInput)));
     }
 
     #[test]
     fn multiple_quality_failure_test() {
         let result = parse_modifiers("ne");
         assert!(result.is_err());
-        assert!(matches!(
-            result,
-            Err(BotError::DiceParsingError(
-                DiceParsingError::InvalidModifiers
-            ))
-        ));
+        assert!(matches!(result, Err(DiceParsingError::InvalidModifiers)));
     }
 
     #[test]
@@ -324,12 +160,7 @@ mod tests {
 
         let result = parse_modifiers("s3q");
         assert!(result.is_err());
-        assert!(matches!(
-            result,
-            Err(BotError::DiceParsingError(
-                DiceParsingError::UnconsumedInput
-            ))
-        ));
+        assert!(matches!(result, Err(DiceParsingError::UnconsumedInput)));
     }
 
     #[test]
@@ -370,6 +201,7 @@ mod tests {
 
     #[test]
     fn dice_pool_complex_expression_test() {
+        use crate::parser::*;
         let modifiers = DicePoolModifiers::custom(DicePoolQuality::Rote, 3);
         let amounts = vec![
             Amount {
